refactor and fix code to work when deployed

+++ db-api/src/main/kotlin/com/vsharkovski/dbpaperapi/service/TempDirectoryService.kt
@@ -4,26 +4,35 @@ import org.springframework.beans.factory.annotation.Value
 import org.springframework.stereotype.Service
 import org.springframework.util.FileSystemUtils
 import java.nio.file.Files
+import java.nio.file.Path
 import java.nio.file.Paths
 import javax.annotation.PostConstruct
+import kotlin.io.path.exists
 
 @Service
 class TempDirectoryService {
     @Value("\${temp.path}")
-    val tempDirectoryPath = ""
+    val tempDirectoryPathString = ""
+
+    lateinit var tempDirectoryPath: Path
 
     @PostConstruct
     fun init() {
         // Create the directory if missing.
-        Files.createDirectories(Paths.get(tempDirectoryPath))
+        tempDirectoryPath = Paths.get(tempDirectoryPathString)
+        Files.createDirectories(tempDirectoryPath)
 
         // Clear it.
-        clearDirectory()
+        clearDirectory(tempDirectoryPath)
     }
 
-    fun clearDirectory() {
+    fun clearDirectory(directory: Path) {
+        if (!directory.exists()) {
+            return
+        }
+
         // Delete all subdirectories and files.
-        Files.newDirectoryStream(Paths.get(tempDirectoryPath)).use { children ->
+        Files.newDirectoryStream(directory).use { children ->
             children.forEach {
                 FileSystemUtils.deleteRecursively(it)
             }
